+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - feat: add command `miactl marketplace list-versions`
 
+### Fixed
+
+- help text of version command
+
 ## [0.9.0] - 2023-11-15
 
 ### Added