+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - help text of version command
+- marketplace apply: silently ignore files with invalid extension rather than throwing error
 
 ## [0.9.0] - 2023-11-15